♻️ create content page for root competencies that only loads when user has access - graphql moved but has not been changed
